Move LateContext methods to librustc_lint.
